# backend/server/gen.go
//go:build go1.22

// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for PasswordGrantRequestGrantType.
const (
	Password PasswordGrantRequestGrantType = "password"
)

// Defines values for Role.
const (
	Admin Role = "admin"
)

// AppError defines model for AppError.
type AppError struct {
	ApiMessage *string                 `json:"api_message,omitempty"`
	Code       string                  `json:"code"`
	StatusCode int32                   `json:"status_code"`
	Values     *map[string]interface{} `json:"values,omitempty"`
}

// Content defines model for Content.
type Content struct {
	// CreatedAt Creation timestamp
	CreatedAt time.Time `json:"created_at"`

	// Data Content data
	Data string `json:"data"`

	// Id Content unique identifier
	Id openapi_types.UUID `json:"id"`

	// OwnerId ID of the user who owns this content
	OwnerId openapi_types.UUID `json:"owner_id"`

	// Type Type of content
	Type string `json:"type"`

	// UpdatedAt Last update timestamp
	UpdatedAt time.Time `json:"updated_at"`
}

// ContentCreateRequest defines model for ContentCreateRequest.
type ContentCreateRequest struct {
	// Data Content data
	Data string `json:"data"`

	// Type Type of content
	Type string `json:"type"`
}

// ContentUpdateRequest defines model for ContentUpdateRequest.
type ContentUpdateRequest struct {
	// Data Content data
	Data *string `json:"data,omitempty"`

	// Type Type of content
	Type *string `json:"type,omitempty"`
}

// PasswordGrantRequest defines model for PasswordGrantRequest.
type PasswordGrantRequest struct {
	// GrantType The grant type for password authentication
	GrantType PasswordGrantRequestGrantType `json:"grant_type"`

	// Password Password of the user
	Password string `json:"password"`

	// Username Username of the user
	Username string `json:"username"`
}

// PasswordGrantRequestGrantType The grant type for password authentication
type PasswordGrantRequestGrantType string

// QrCode defines model for QrCode.
type QrCode struct {
	// CreatedAt Creation timestamp
	CreatedAt time.Time `json:"created_at"`

	// Id QR code unique identifier
	Id openapi_types.UUID `json:"id"`

	// OwnerId ID of the user who owns this QR code
	OwnerId openapi_types.UUID `json:"owner_id"`

	// Slug Unique slug for accessing this QR code
	Slug string `json:"slug"`

	// UpdatedAt Last update timestamp
	UpdatedAt time.Time `json:"updated_at"`
}

// QrCodeCreateRequest defines model for QrCodeCreateRequest.
type QrCodeCreateRequest struct {
	// Slug Custom slug for the QR code (optional)
	Slug *string `json:"slug,omitempty"`
}

// Role defines model for Role.
type Role string

// Token defines model for Token.
type Token struct {
	// AccessToken The access token for the user
	AccessToken string `json:"access_token"`

	// ExpiresIn The number of seconds until the access token expires
	ExpiresIn int `json:"expires_in"`

	// RefreshToken The refresh token for the user
	RefreshToken string `json:"refresh_token"`

	// TokenType The type of the token (e.g., Bearer)
	TokenType string `json:"token_type"`
}

// User defines model for User.
type User struct {
	// Email Email address of the user
	Email *string `json:"email,omitempty"`

	// Id User ID
	Id string `json:"id"`

	// Metadata Additional metadata about the user
	Metadata *map[string]interface{} `json:"metadata,omitempty"`
	Roles    *[]Role                 `json:"roles,omitempty"`

	// Username Username of the user
	Username string `json:"username"`
}

// Forbidden defines model for Forbidden.
type Forbidden = AppError

// InternalServerError defines model for InternalServerError.
type InternalServerError = AppError

// Unauthorized defines model for Unauthorized.
type Unauthorized = AppError

// CreateContentJSONRequestBody defines body for CreateContent for application/json ContentType.
type CreateContentJSONRequestBody = ContentCreateRequest

// UpdateContentJSONRequestBody defines body for UpdateContent for application/json ContentType.
type UpdateContentJSONRequestBody = ContentUpdateRequest

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody = PasswordGrantRequest

// CreateQrCodeJSONRequestBody defines body for CreateQrCode for application/json ContentType.
type CreateQrCodeJSONRequestBody = QrCodeCreateRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get all content
	// (GET /content)
	GetContentItems(w http.ResponseWriter, r *http.Request)
	// Create new content
	// (POST /content)
	CreateContent(w http.ResponseWriter, r *http.Request)
	// Delete content
	// (DELETE /content/{id})
	DeleteContent(w http.ResponseWriter, r *http.Request, id openapi_types.UUID)
	// Get content by ID
	// (GET /content/{id})
	GetContentById(w http.ResponseWriter, r *http.Request, id openapi_types.UUID)
	// Update content
	// (PUT /content/{id})
	UpdateContent(w http.ResponseWriter, r *http.Request, id openapi_types.UUID)
	// Login with username and password
	// (POST /login)
	LoginUser(w http.ResponseWriter, r *http.Request)
	// Get the current user
	// (GET /me)
	GetCurrentUser(w http.ResponseWriter, r *http.Request)
	// Get all QR codes
	// (GET /qrcodes)
	GetQrCodes(w http.ResponseWriter, r *http.Request)
	// Create a new QR code
	// (POST /qrcodes)
	CreateQrCode(w http.ResponseWriter, r *http.Request)
	// Get a QR code by Slug
	// (GET /qrcodes/by-slug/{slug})
	GetQrCodeBySlug(w http.ResponseWriter, r *http.Request, slug string)
	// Get a QR code by ID
	// (GET /qrcodes/{id})
	GetQrCodeById(w http.ResponseWriter, r *http.Request, id openapi_types.UUID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetContentItems operation middleware
func (siw *ServerInterfaceWrapper) GetContentItems(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetContentItems(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateContent operation middleware
func (siw *ServerInterfaceWrapper) CreateContent(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateContent(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteContent operation middleware
func (siw *ServerInterfaceWrapper) DeleteContent(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteContent(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetContentById operation middleware
func (siw *ServerInterfaceWrapper) GetContentById(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetContentById(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateContent operation middleware
func (siw *ServerInterfaceWrapper) UpdateContent(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateContent(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// LoginUser operation middleware
func (siw *ServerInterfaceWrapper) LoginUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.LoginUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetCurrentUser operation middleware
func (siw *ServerInterfaceWrapper) GetCurrentUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetCurrentUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetQrCodes operation middleware
func (siw *ServerInterfaceWrapper) GetQrCodes(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetQrCodes(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateQrCode operation middleware
func (siw *ServerInterfaceWrapper) CreateQrCode(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateQrCode(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetQrCodeBySlug operation middleware
func (siw *ServerInterfaceWrapper) GetQrCodeBySlug(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameterWithOptions("simple", "slug", r.PathValue("slug"), &slug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "slug", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetQrCodeBySlug(w, r, slug)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetQrCodeById operation middleware
func (siw *ServerInterfaceWrapper) GetQrCodeById(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetQrCodeById(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/content", wrapper.GetContentItems)
	m.HandleFunc("POST "+options.BaseURL+"/content", wrapper.CreateContent)
	m.HandleFunc("DELETE "+options.BaseURL+"/content/{id}", wrapper.DeleteContent)
	m.HandleFunc("GET "+options.BaseURL+"/content/{id}", wrapper.GetContentById)
	m.HandleFunc("PUT "+options.BaseURL+"/content/{id}", wrapper.UpdateContent)
	m.HandleFunc("POST "+options.BaseURL+"/login", wrapper.LoginUser)
	m.HandleFunc("GET "+options.BaseURL+"/me", wrapper.GetCurrentUser)
	m.HandleFunc("GET "+options.BaseURL+"/qrcodes", wrapper.GetQrCodes)
	m.HandleFunc("POST "+options.BaseURL+"/qrcodes", wrapper.CreateQrCode)
	m.HandleFunc("GET "+options.BaseURL+"/qrcodes/by-slug/{slug}", wrapper.GetQrCodeBySlug)
	m.HandleFunc("GET "+options.BaseURL+"/qrcodes/{id}", wrapper.GetQrCodeById)

	return m
}